moved Slider in to separate component
